[#154] Refactored Comparator component Closes #154 * Added ComparatorContainer that renders Comparator
