+++ frontend/src/components/App.js
@@ -3,10 +3,10 @@ import Header from "./Header.js";
 import { Switch, Route } from "react-router-dom";
 import Welcome from "./Welcome.js";
 import Library from "./Library.js";
-import Comparator from "./Comparator.js";
 import ParkingSpaceContainer from "../containers/ParkingSpacesContainer.js";
 import OutsideSensorsContainer from "../containers/OutsideSensorsContainer.js"
 import SingleSensorPageContainer from "../containers/SingleSensorPageContainer.js";
+import ComparatorContainer from "../containers/ComparatorContainer.js";
 
 function App() {
   return (
@@ -19,7 +19,7 @@ function App() {
         <Route path="/outside-sensors" component={OutsideSensorsContainer} exact/>
         <Route path="/parking-spaces" component={ParkingSpaceContainer} />
         <Route path="/sensor" component={SingleSensorPageContainer}/>
-        <Route path="/compare" component={Comparator} />
+        <Route path="/compare" component={ComparatorContainer} />
         <Route component={Error} />
       </Switch>
     </main>
